Suppress a few warnings.
Switch to the support library from AndroidX to remove usages of the
deprecated getFragmentManager() API and android.app.Fragment class.

Update the build tools to version 30.0.1 and the SDK version to 30.

Update the target API Level to 28, the last version allowing
execute permission for files stored within the app's home directory.