Switch to a more recent version of Erlang, release 23.1.5
Update both the 32-bit and 64-bit variants of the Erlang runtime
to release 23.1.5 built with OpenSSL 1.1.1i (linked statically).

Include the crypto and ssl applications, this fixes #4. For even more
applications, the full Erlang releases for Android are published here:
https://github.com/JeromeDeBretagne/erlang-otp-releases-for-android

Switch to the Jinterface package (OtpErlang.jar) from Erlang 23.1.5 too.