Switch to a more generic approach to copy the Erlang code
Define an 'ebin' subdirectory within the app assets where to drop all
the .beam files and extract them in an 'ebin' directory in the app data.
This removes the need to hard-code the name of the file(s) to extract.

Move hello_jinterface.beam in this new asset 'ebin' subdirectory.